Compare system state against data

Compare current system state against data. This is only supported for systems that have 'compare_data' support (i.e., the system definition includes a compare_data method). The current state in the system (dust_system_state) is compared against the data provided as data.

Usage

dust_system_compare_data(sys, data)

Arguments

sys

A dust_system object

data

The data to compare against. If the system is ungrouped then data is a list with elements corresponding to whatever your system requires. If your system is grouped, this should be a list with as many elements as your system has groups, with each element corresponding to the data your system requires.

Value

A numeric vector with as many elements as your system has groups, corresponding to the log-likelihood of the data for each group.