Create a dust object

Create a dust system object from a system generator. This allocates a system and sets an initial set of parameters. Once created you can use other dust functions to interact with it.

Usage

dust_system_create(
  generator,
  pars,
  n_particles,
  n_groups = 0,
  time = 0,
  dt = NULL,
  ode_control = NULL,
  seed = NULL,
  deterministic = FALSE
)

Arguments

generator

A system generator object, with class dust_system_generator

pars

A list of parameters. The format of this will depend on the system. If n_groups is 1 or more, then this must be a list of length n_groups where each element is a list of parameters for your system.

n_particles

The number of particles to create.

n_groups

Optionally, the number of parameter groups

time

The initial time, defaults to 0

dt

The time step for discrete time systems, defaults to 1 if not given. It is an error to provide a non-NULL argument with continuous-time systems.

ode_control

The ODE integration control for continuous time systems. Defaults to the default return of dust_ode_control. It is an error to provide this with discrete-time systems.

seed

Optionally, a seed. Otherwise we respond to R's RNG seed on initialisation.

deterministic

Logical, indicating if the system should be allocated in deterministic mode.

Value

A dust_system object, with opaque format.