Create a system generator

Create a system generator. This function is not intended to be used by users directly, but will be called from packages built using dust_package

Usage

dust_system_generator(
  name,
  time_type,
  default_dt,
  env = parent.env(parent.frame())
)

Arguments

name

The name of the generator

time_type

The time type (discrete or continuous). Using the wrong time here will lead to crashes or failure to create the generator.

default_dt

The default value for dt on initialisation

env

The environment where the generator is defined.

Value

A dust_system_generator object

Examples

# This creates the "sir" generator
dust_system_generator("sir", "discrete", 1, asNamespace("dust2"))
#> 
#> ── <dust_system_generator: sir> ────────────────────────────────────────────────
#> ℹ Use 'dust2::dust_system_create()' to create a system with this generator
#> ℹ This system has 'compare_data' support
#> ℹ This system runs in discrete time with a default dt of 1

# This is the same code as in "dust2:::sir", except there we find
# the correct environment automatically
dust2:::sir
#> function () 
#> {
#>     dust_system_generator("sir", "discrete", 1)
#> }
#> <bytecode: 0x5581be0f4100>
#> <environment: namespace:dust2>