Fetch system internals

Return internal data from the system. This is intended for debugging only, and all formats are subject to change.

Usage

dust_system_internals(sys)

Arguments

sys

A dust_system object

Value

If sys is a discrete-time system, this function returns NULL, as no internal data is stored. Otherwise, for a continuous-time system we return a data.frame of statistics with one row per particle. Most of the columns are simple integers or numeric values, but dydt (the current derivative of the target function with respect to time) and step_times (times that the solver has stopped at, if debug_record_step_times is in dust_ode_control was set to TRUE) will be a list of columns, each element of which is a numeric vector.