Set system time

Set time into the system. This updates the time to the provided value but does not affect the state. You may want to call dust_system_set_state or dust_system_set_state_initial after calling this.

Usage

dust_system_set_time(sys, time)

Arguments

sys

A dust_system object

time

The time to set. Currently this must be an integer-like value, but in future we will allow setting to any multiple of dt.

Value

Nothing, called for side effects only