Unpack state

Unpack state. You will see state come out of dust2 systems in several places, for example dust_system_state, but it will typically be an unstructured vector with no names; this is not very useful! However, your model knows what each element, or group of elements "means". You can unpack your state from this unstructured array into a named list using this function. The same idea applies to the higher-dimensional arrays that you might get if your system has multiple particles, multiple parameter groups or has been run for multiple time steps.

Usage

dust_unpack_state(obj, state)

dust_unpack_index(obj)

Arguments

obj

A dust_system object (from dust_system_create) or dust_likelihood object (from dust_filter_create or dust_unfilter_create).

state

A state vector, matrix or array. This might have come from dust_system_state, dust_likelihood_last_history, or dust_likelihood_last_state.

Value

A named list, where each element corresponds to a logical compartment.

See also

Examples

sys <- dust_system_create(dust2:::sir(), list(), n_particles = 10, dt = 0.25)
dust_system_set_state_initial(sys)
t <- seq(0, 100, by = 5)
y <- dust_system_simulate(sys, t)
# The result here is a 5 x 10 x 21 matrix: 5 states by 10 particles by
# 21 times.
dim(y)
#> [1]  5 10 21

# The 10 particles and 21 times (following t) are simple enough, but
# what are our 5 compartments?

# You can use dust_unpack_state() to reshape your output as a
# list:
dust_unpack_state(sys, y)
#> $S
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]  990  978  964  949  930  912  897  875  855   828   780   738   678
#>  [2,]  990  984  974  955  938  890  832  739  655   567   462   379   341
#>  [3,]  990  981  970  951  922  881  826  741  657   576   484   426   359
#>  [4,]  990  976  955  934  903  844  774  678  576   497   416   359   309
#>  [5,]  990  988  980  970  932  893  843  783  711   599   501   418   361
#>  [6,]  990  981  964  942  906  853  795  689  588   516   444   376   318
#>  [7,]  990  987  982  976  963  932  894  855  797   708   614   542   462
#>  [8,]  990  979  968  942  906  848  787  693  587   497   408   342   302
#>  [9,]  990  983  963  915  852  770  659  559  493   413   350   313   285
#> [10,]  990  982  966  942  908  869  817  752  690   619   544   478   418
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21]
#>  [1,]   603   538   473   422   393   362   337   317
#>  [2,]   304   274   249   235   227   221   215   214
#>  [3,]   327   292   269   252   241   227   219   211
#>  [4,]   281   265   248   239   231   230   230   227
#>  [5,]   318   296   276   263   245   235   230   225
#>  [6,]   284   264   245   235   222   210   205   202
#>  [7,]   406   359   326   289   266   258   254   249
#>  [8,]   279   254   246   235   225   217   211   208
#>  [9,]   267   243   224   220   216   215   214   213
#> [10,]   373   329   303   275   263   253   244   238
#> 
#> $I
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   10   15   18   23   28   29   27   36   36    47    62    72   100
#>  [2,]   10   13   11   25   24   57   73  123  147   172   183   180   140
#>  [3,]   10   14   16   24   40   56   83  117  142   145   170   151   138
#>  [4,]   10   20   31   36   43   78  107  145  170   168   166   140   123
#>  [5,]   10    5   10   17   43   54   78   90  126   171   187   174   144
#>  [6,]   10   18   24   31   47   71   96  138  162   152   151   142   139
#>  [7,]   10    4    7   11   18   35   49   62   88   130   148   143   157
#>  [8,]   10   17   17   29   50   81   97  147  170   181   177   150   126
#>  [9,]   10   12   21   47   81  118  149  164  153   160   144   126    96
#> [10,]   10   12   21   33   45   63   83  107  108   120   130   126   129
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21]
#>  [1,]   119   129   131   117    94    89    73    63
#>  [2,]   114    98    82    67    42    32    26    16
#>  [3,]   101    90    67    51    45    37    27    21
#>  [4,]    93    60    46    35    20    13     9     7
#>  [5,]   123    85    72    54    46    36    28    23
#>  [6,]   115    81    65    43    43    38    27    14
#>  [7,]   143   120   112    86    69    41    24    16
#>  [8,]   107    78    52    41    37    30    23    14
#>  [9,]    66    58    45    25    20    17    11     8
#> [10,]   110   103    90    72    44    34    25    20
#> 
#> $R
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    7   18   28   42   59   76   89  109   125   158   190   222
#>  [2,]    0    3   15   20   38   53   95  138  198   261   355   441   519
#>  [3,]    0    5   14   25   38   63   91  142  201   279   346   423   503
#>  [4,]    0    4   14   30   54   78  119  177  254   335   418   501   568
#>  [5,]    0    7   10   13   25   53   79  127  163   230   312   408   495
#>  [6,]    0    1   12   27   47   76  109  173  250   332   405   482   543
#>  [7,]    0    9   11   13   19   33   57   83  115   162   238   315   381
#>  [8,]    0    4   15   29   44   71  116  160  243   322   415   508   572
#>  [9,]    0    5   16   38   67  112  192  277  354   427   506   561   619
#> [10,]    0    6   13   25   47   68  100  141  202   261   326   396   453
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21]
#>  [1,]   278   333   396   461   513   549   590   620
#>  [2,]   582   628   669   698   731   747   759   770
#>  [3,]   572   618   664   697   714   736   754   768
#>  [4,]   626   675   706   726   749   757   761   766
#>  [5,]   559   619   652   683   709   729   742   752
#>  [6,]   601   655   690   722   735   752   768   784
#>  [7,]   451   521   562   625   665   701   722   735
#>  [8,]   614   668   702   724   738   753   766   778
#>  [9,]   667   699   731   755   764   768   775   779
#> [10,]   517   568   607   653   693   713   731   742
#> 
#> $cases_cumul
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0   12   26   41   60   78   93  115  135   162   210   252   312
#>  [2,]    0    6   16   35   52  100  158  251  335   423   528   611   649
#>  [3,]    0    9   20   39   68  109  164  249  333   414   506   564   631
#>  [4,]    0   14   35   56   87  146  216  312  414   493   574   631   681
#>  [5,]    0    2   10   20   58   97  147  207  279   391   489   572   629
#>  [6,]    0    9   26   48   84  137  195  301  402   474   546   614   672
#>  [7,]    0    3    8   14   27   58   96  135  193   282   376   448   528
#>  [8,]    0   11   22   48   84  142  203  297  403   493   582   648   688
#>  [9,]    0    7   27   75  138  220  331  431  497   577   640   677   705
#> [10,]    0    8   24   48   82  121  173  238  300   371   446   512   572
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21]
#>  [1,]   387   452   517   568   597   628   653   673
#>  [2,]   686   716   741   755   763   769   775   776
#>  [3,]   663   698   721   738   749   763   771   779
#>  [4,]   709   725   742   751   759   760   760   763
#>  [5,]   672   694   714   727   745   755   760   765
#>  [6,]   706   726   745   755   768   780   785   788
#>  [7,]   584   631   664   701   724   732   736   741
#>  [8,]   711   736   744   755   765   773   779   782
#>  [9,]   723   747   766   770   774   775   776   777
#> [10,]   617   661   687   715   727   737   746   752
#> 
#> $cases_inc
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    4    6    6    4    7    4    3    4     7    13     9    10
#>  [2,]    0    4    4    8    2   12    9   27   13    21    14    18     9
#>  [3,]    0    3    3    2   11   10   11   14   18    21    18    11    10
#>  [4,]    0    4    4    4    8   23   18   15   18    16    13     9    13
#>  [5,]    0    0    3    6    9    9    6   15   19    20    29    14     6
#>  [6,]    0    4    3    7    9    6   15   23   26    19    15    14    11
#>  [7,]    0    0    0    0    3    5   11    8   13    19    20    13    23
#>  [8,]    0    1    2    8    8   18   14   18   21     9    20     8     7
#>  [9,]    0    2    1   13   14   21   20   16   10    10    10    10     6
#> [10,]    0    1    1    5   11    7    9    8   12    15    11     7    11
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21]
#>  [1,]    20     9    11    12     7     4     5     6
#>  [2,]     8    11     7     2     2     0     2     0
#>  [3,]     7     6     7     2     1     2     2     1
#>  [4,]     5     4     2     1     1     0     0     1
#>  [5,]     8     4     2     3     2     3     1     1
#>  [6,]     4     4     4     4     3     2     0     1
#>  [7,]    11     8    13     6     3     1     1     0
#>  [8,]     8     5     5     3     3     0     2     0
#>  [9,]     3     4     7     1     0     0     0     0
#> [10,]    10     9     5     6     2     1     0     1
#> 

# Here, the list is named following the compartments (S, I, R,
# etc) and is a 10 x 21 matrix (i.e., the remaining dimensions
# from y)

# We could apply this to the final state, which converts a 5 x 10
# matrix of state into a 5 element list of vectors, each with
# length 10:
s <- dust_system_state(sys)
dim(s)
#> [1]  5 10
dust_unpack_state(sys, s)
#> $S
#>  [1] 317 214 211 227 225 202 249 208 213 238
#> 
#> $I
#>  [1] 63 16 21  7 23 14 16 14  8 20
#> 
#> $R
#>  [1] 620 770 768 766 752 784 735 778 779 742
#> 
#> $cases_cumul
#>  [1] 673 776 779 763 765 788 741 782 777 752
#> 
#> $cases_inc
#>  [1] 6 0 1 1 1 1 0 0 0 1
#> 

# If you need more control, you can use 'dust_unpack_index' to map
# names to positions within the state dimension of this array
dust_unpack_index(sys)
#> $S
#> [1] 1
#> 
#> $I
#> [1] 2
#> 
#> $R
#> [1] 3
#> 
#> $cases_cumul
#> [1] 4
#> 
#> $cases_inc
#> [1] 5
#>