Fetch system internals

Return internal data from the system. This is intended for debugging only, and all formats are subject to change.

Usage

dust_system_internals(
  sys,
  include_coefficients = FALSE,
  include_history = FALSE
)

Arguments

sys

A dust_system object

include_coefficients

Boolean, indicating if interpolation coefficients should be included in the output. These are intentionally undocumented for now.

include_history

Boolean, also undocumented.

Value

If sys is a discrete-time system, this function returns NULL, as no internal data is stored. Otherwise, for a continuous-time system we return a data.frame of statistics with one row per particle. Most of the columns are simple integers or numeric values, but dydt (the current derivative of the target function with respect to time) and step_times (times that the solver has stopped at, if debug_record_step_times is in dust_ode_control was set to TRUE) will be a list of columns, each element of which is a numeric vector. If include_coefficients is TRUE, the coefficients column exists and holds a list of coefficients (the structure of these may change over time, too).